+++ src/jawl_voicecompanion/__init__.py
@@ -4,6 +4,7 @@ from .gateway import TextGateway
 from .hostos_policy import HostOSPolicy
 from .hostos_tools import HostOSExecutor, ToolSpec
 from .browser_adapter import BrowserAdapter
+from .approvals import ApprovalStore
 from .models import AccessLevel, RiskClass, ToolRequest
 from .arbiter import TurnArbiter, TurnPriority
 
@@ -12,6 +13,7 @@ __all__ = [
     "HostOSPolicy",
     "HostOSExecutor",
     "BrowserAdapter",
+    "ApprovalStore",
     "RiskClass",
     "TextGateway",
     "ToolRequest",